Refactor - moved WallOfResurgenceToken to shared area
