+++ Mage.Sets/src/mage/sets/oathofthegatewatch/CycloneSire.java
@@ -41,6 +41,7 @@ import mage.constants.Duration;
 import mage.constants.Rarity;
 import mage.counters.CounterType;
 import mage.filter.common.FilterControlledLandPermanent;
+import mage.game.permanent.token.WallOfResurgenceToken;
 import mage.target.common.TargetControlledPermanent;
 
 /**
