+++ Mage.Sets/src/mage/sets/oathofthegatewatch/WallOfResurgence.java
@@ -35,14 +35,13 @@ import mage.abilities.effects.Effect;
 import mage.abilities.effects.common.continuous.BecomesCreatureTargetEffect;
 import mage.abilities.effects.common.counter.AddCountersTargetEffect;
 import mage.abilities.keyword.DefenderAbility;
-import mage.abilities.keyword.HasteAbility;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Duration;
 import mage.constants.Rarity;
 import mage.counters.CounterType;
 import mage.filter.common.FilterControlledLandPermanent;
-import mage.game.permanent.token.Token;
+import mage.game.permanent.token.WallOfResurgenceToken;
 import mage.target.common.TargetControlledPermanent;
 
 /**
@@ -76,19 +75,5 @@ public class WallOfResurgence extends CardImpl {
     @Override
     public WallOfResurgence copy() {
         return new WallOfResurgence(this);
-    }
-}
-
-class WallOfResurgenceToken extends Token {
-
-    public WallOfResurgenceToken() {
-        super("", "0/0 Elemental creature with haste");
-        this.cardType.add(CardType.CREATURE);
-
-        this.subtype.add("Elemental");
-        this.power = new MageInt(0);
-        this.toughness = new MageInt(0);
-
-        this.addAbility(HasteAbility.getInstance());
     }
 }
