+++ Mage/src/main/java/mage/game/permanent/token/WallOfResurgenceToken.java
@@ -1,0 +1,19 @@
+package mage.game.permanent.token;
+
+import mage.MageInt;
+import mage.abilities.keyword.HasteAbility;
+import mage.constants.CardType;
+
+public class WallOfResurgenceToken extends Token {
+
+    public WallOfResurgenceToken() {
+        super("", "0/0 Elemental creature with haste");
+        this.cardType.add(CardType.CREATURE);
+
+        this.subtype.add("Elemental");
+        this.power = new MageInt(0);
+        this.toughness = new MageInt(0);
+
+        this.addAbility(HasteAbility.getInstance());
+    }
+}
